requirements.txt
deepagents
langsmith
tavily-python
python-dotenv
pandas
scipy
matplotlib
seaborn
statsmodels
requests
openpyxl
pyalex
biopython
